tiffly-admin/requirements.txt
addressable==1.4.2
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.7.5
aniso8601==9.0.1
apt-wrapper==1.18
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.6.0
bcrypt==3.2.2
blinker==1.6.2
boto3==1.20.46
botocore==1.23.46
cachelib==0.9.0
cachetools==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
Flask==2.0.1
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
frozenlist==1.3.3
google-analytics-data==0.15.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==1.1.2
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
httplib2==0.21.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
inspect-it==0.3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==4.2.1
keyring==4.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.18.0
multidict==6.0.4
numpy==1.22.1
oauth2client==4.1.3
packaging==21.3
pickle5==0.0.11
Pillow==9.0.0
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyfcm==1.5.4
PyJWT==2.6.0
pylev==1.4.0
pymongo==4.0
pyparsing==3.0.9
pyrsistent==0.18.0
pysolr==3.9.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.27.1
rsa==3.1.4
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
twilio==8.2.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
voluptuous==0.12.2
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==2.3.3
yarl==1.9.2
zipp==3.6.0
